+++ data_source_disk.go
@@ -70,6 +70,8 @@ func dataSourceDiskRead(d *schema.ResourceData, m interface{}) error {
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the disk - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	disk := DiskBody{}