+++ data_source_disks.go
@@ -71,6 +71,8 @@ func dataSourceDisksRead(d *schema.ResourceData, m interface{}) error {
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the disks - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	disks := DiskListBody{}