+++ data_source_firewall_rule.go
@@ -84,6 +84,8 @@ func dataSourceFirewallRuleRead(d *schema.ResourceData, m interface{}) error {
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the firewall rule - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	firewallRule := FirewallRuleBody{}