+++ data_source_ip_addresses.go
@@ -78,6 +78,8 @@ func dataSourceIPAddressesRead(d *schema.ResourceData, m interface{}) error {
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the IP addresses - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	ipAddresses := IPAddressListBody{}