+++ data_source_network_interface.go
@@ -139,6 +139,8 @@ func dataSourceNetworkInterfaceRead(d *schema.ResourceData, m interface{}) error
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the network interface - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	networkInterface := NetworkInterfaceBody{}