+++ data_source_servers.go
@@ -123,6 +123,8 @@ func dataSourceServersRead(d *schema.ResourceData, m interface{}) error {
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the servers - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	list := make(ServerListBody, 0)