+++ data_source_templates.go
@@ -81,6 +81,8 @@ func dataSourceTemplatesRead(d *schema.ResourceData, m interface{}) error {
 
 	if resErr != nil {
 		return resErr
+	} else if res.StatusCode != 200 {
+		return fmt.Errorf("Failed to read the information about the templates - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	list := make(TemplateListBody, 0)