+++ CHANGELOG.md
@@ -14,3 +14,4 @@ FEATURES:
 * **New Data Source:** `clouddk_server`
 * **New Data Source:** `clouddk_servers`
 * **New Data Source:** `clouddk_templates`
+* **New Resource:** `clouddk_server`